Fold the water-tight test files into their home suites
The water-tight tests split by layer, matching the existing suites, so merge
them in rather than keep standalone files:

- test_vbd_water_tight_contact.py -> test_solver_vbd.py: the SolverVBD
  consumption tests (class TestVBDWaterTightContact). test_solver_vbd now runs 60.
- test_water_tight_soft_contact.py -> test_collision_pipeline.py: the SDF
  detection/pipeline tests (class TestWaterTightSoftContact). Its
  get_test_devices() set is kept as `soft_devices` so it does not collide with
  the file's cuda-basic `devices`. test_collision_pipeline now runs 133.

Test-only move; no test logic changed (imports merged, one stale "kind"
docstring corrected).
